# Fix attention caching to make transcription run 30% faster