For a `dict.get(a, f())` call, `f()` is run even when the `dict` already has `a`. This is causing the bug in attention caching logic.
So making this change fixes the caching logic, and gets a sweet (expected) speedup of close to 30%. 

For a close to 5 min test audio, on an Nvidia Quadro RTX 8000 gpu:
* before this fix, the `large` model took 102.8s, and after this fix it takes 68.9s
* before this fix, the `medium.en` model took 55.1s, and after this fix it takes 39.7s