@janyf 2022-10-19
confirming that this really speed up the process, nice catch ! 

large model slovak language , 16min file on GTX1080 Ti
before change: ~22min
after change: ~12min