@akashmjn 2022-10-19
Can repro on my end as well

[17:38 wav file](https://drive.google.com/file/d/1a6jPRu8nF84PSQesDz-NKsDbheatyuw8/view?usp=sharing)
On Quadro RTX 5000  (with `--condition_on_previous_text=True`, `small.en` model)

|                     | Runtime |
|---------------------|---------|
| Original            | 01:42   |
| With kv cache fixed | 01:24   |

Really elegant catch! 🙌 learnt something about Python today 🙂