@nlgtuankiet 2022-10-20
@vickianand @ritheshkumar95 
Thank you very much!
I was getting around 0.88-1.1x for large models, now it 1.6x