@rjwilmsi 2022-10-20
Wow, this is great news, thank you to all involved. Title says 30% faster on GPU but the speed up is much more on CPU, at least for me.

Ryzen 5 4500U / 6 core laptop CPU. 6m30s youtube video, English (https://www.youtube.com/watch?v=GFu64hnqzVo):

| Model         | Before PR     | After PR |
|--------------|-----------|------------|
| tiny.en | 6m19s      | 2m9s        |
| base.en      | 15m39s  | 4m29s       |
| small.en    | 60m45s | 11m18s   |